glslang -> Top IR: fully handle all global initializers.
This includes declarations with initializers that can appear between functions. These can include the ?: operator, which introduces flow control. Semantically, this code is to execute as the first code in main(). That means, after main() is translated, more global declarations with initializers could be seen and inserted early in main(), and even add basic blocks.