Building: On *nix machines, set LLVMINCLUDES and LLVMLIBS out of the install directory.  From Lei Zhang <antiagainst@google.com>
This essentially removes the requirement to install LLVM 3.4 to build/install on *nix machines.

We find this patch useful if we want to build LLVM 3.4 as a custom target on which LunarGLASS depends. With this patch, make for LLVM 3.4 will not go over every directory trying to find new stuff to install. It saves build time.

Kindly remind that both LLVMINCLUDES and LLVMLIBS are cached variables, so please purge the build tree to test out.